[PW-7459] Remove integration tests (#424)
* Remove integration tests

* Replace the integration tests with unit tests

* Fix a small issue from sonal cloud

* Fix format error

* Update comment
