+++ tests/Resources/Management/get-merchants-success.json
@@ -1,0 +1,81 @@
+{
+  "_links": {
+    "first": {
+      "href": "https://management-test.adyen.com/v1/merchants?pageNumber=1&pageSize=10"
+    },
+    "last": {
+      "href": "https://management-test.adyen.com/v1/merchants?pageNumber=1&pageSize=10"
+    },
+    "self": {
+      "href": "https://management-test.adyen.com/v1/merchants?pageNumber=1&pageSize=10"
+    }
+  },
+  "itemsTotal": 2,
+  "pagesTotal": 1,
+  "data": [
+    {
+      "id": "TestMerchantAccount1",
+      "name": "TestMerchantAccount1",
+      "companyId": "Demo",
+      "captureDelay": "manual",
+      "defaultShopperInteraction": "Ecommerce",
+      "status": "Active",
+      "shopWebAddress": "https://TestMerchantAccount1.store/Sites-Site",
+      "merchantCity": "Amsterdam",
+      "primarySettlementCurrency": "EUR",
+      "dataCenters": [
+        {
+          "name": "default",
+          "livePrefix": ""
+        }
+      ],
+      "_links": {
+        "self": {
+          "href": "https://management-test.adyen.com/v1/merchants/TestMerchantAccount1"
+        },
+        "apiCredentials": {
+          "href": "https://management-test.adyen.com/v1/merchants/TestMerchantAccount1/apiCredentials"
+        },
+        "users": {
+          "href": "https://management-test.adyen.com/v1/merchants/TestMerchantAccount1/users"
+        },
+        "webhooks": {
+          "href": "https://management-test.adyen.com/v1/merchants/TestMerchantAccount1/webhooks"
+        }
+      },
+      "description": "TestMerchantAccount1"
+    },
+    {
+      "id": "TestMerchantAccount2",
+      "name": "TestMerchantAccount2",
+      "companyId": "Demo",
+      "captureDelay": "immediate",
+      "defaultShopperInteraction": "Ecommerce",
+      "status": "Active",
+      "shopWebAddress": "https://TestMerchantAccount1.store/Sites-Site",
+      "merchantCity": "Amsterdam",
+      "primarySettlementCurrency": "EUR",
+      "dataCenters": [
+        {
+          "name": "default",
+          "livePrefix": ""
+        }
+      ],
+      "_links": {
+        "self": {
+          "href": "https://management-test.adyen.com/v1/merchants/TestMerchantAccount2"
+        },
+        "apiCredentials": {
+          "href": "https://management-test.adyen.com/v1/merchants/TestMerchantAccount2/apiCredentials"
+        },
+        "users": {
+          "href": "https://management-test.adyen.com/v1/merchants/TestMerchantAccount2/users"
+        },
+        "webhooks": {
+          "href": "https://management-test.adyen.com/v1/merchants/TestMerchantAccount2/webhooks"
+        }
+      },
+      "description": "TestMerchantAccount2"
+    }
+  ]
+}
