+++ tests/Resources/Management/get-payment-method-settings-success.json
@@ -1,0 +1,10 @@
+{
+  "id": "PM3227C223224K5FH84M8CBNH",
+  "type": "visa",
+  "countries": [
+    "US"
+  ],
+  "currencies": [
+    "USD"
+  ]
+}
