+++ tests/Resources/Management/post-companies-webhook-success.json
@@ -1,0 +1,95 @@
+{
+  "id" : "S2-31433949437F",
+  "type" : "standard",
+  "url" : "YOUR_WEBHOOK_URL",
+  "description" : "Webhook for YOUR_MERCHANT_ACCOUNT - YOUR_MERCHANT_CODE",
+  "username" : "myuser",
+  "hasPassword" : true,
+  "active" : true,
+  "hasError" : false,
+  "sslVersion" : "TLSv1.2",
+  "communicationFormat" : "json",
+  "acceptsExpiredCertificate" : false,
+  "acceptsSelfSignedCertificate" : true,
+  "acceptsUntrustedRootCertificate" : true,
+  "certificateAlias" : "signed-test.adyen.com_2022",
+  "populateSoapActionHeader" : false,
+  "additionalSettings" : {
+    "properties" : {
+      "includePosTerminalInfo" : false,
+      "includeARN" : false,
+      "includePosDetails" : false,
+      "includeCardInfoForRecurringContractNotification" : false,
+      "includeRiskData" : false,
+      "includeRiskExperimentReference" : false,
+      "includeSoapSecurityHeader" : false,
+      "includeContactlessWalletTokenInformation" : false,
+      "includeAcquirerReference" : false,
+      "includeRiskProfileReference" : false,
+      "includeOriginalMerchantReferenceCancelOrRefundNotification" : false,
+      "includeNfcTokenInformation" : false,
+      "includeSubvariant" : false,
+      "includeThreeDSVersion" : false,
+      "includeInstallmentsInfo" : false,
+      "includeAliasInfo" : false,
+      "includeShopperCountry" : false,
+      "includeRawThreeDSecureResult" : false,
+      "includeAirlineData" : false,
+      "includeGrossCurrencyChargebackDetails" : false,
+      "includeThreeDSecureResult" : false,
+      "includeMetadataIn3DSecurePaymentNotification" : false,
+      "includeOriginalReferenceForChargebackReversed" : false,
+      "addAcquirerResult" : false,
+      "includeDeliveryAddress" : false,
+      "includeRetryAttempts" : false,
+      "includeExtraCosts" : false,
+      "includeCardHolderName" : false,
+      "includeShopperDetail" : false,
+      "includeBankAccountDetails" : false,
+      "includeMandateDetails" : false,
+      "includeAuthAmountForDynamicZeroAuth" : false,
+      "includeIssuerCountry" : false,
+      "includeAcquirerErrorDetails" : false,
+      "includeCoBrandedWith" : false,
+      "includeShopperInteraction" : false,
+      "includeDeviceAndBrowserInfo" : false,
+      "addRawAcquirerResult" : false,
+      "includeCardBin" : false,
+      "includeFundingSource" : false,
+      "includeThreeDS2ChallengeInformation" : false,
+      "includeRiskProfile" : false,
+      "includeRealtimeAccountUpdaterStatus" : false,
+      "includeDunningProjectData" : false,
+      "includePaymentResultInOrderClosedNotification" : false,
+      "includeCardBinDetails" : false,
+      "includeNotesInManualReviewNotifications" : false,
+      "includeZeroAuthFlag" : false,
+      "addCaptureReferenceToDisputeNotification" : false,
+      "includePayPalDetails" : false,
+      "includeRawThreeDSecureDetailsResult" : false,
+      "includeBankVerificationResults" : false,
+      "includeCaptureDelayHours" : false,
+      "addPaymentAccountReference" : false,
+      "includePayULatamDetails" : false,
+      "includeStore" : false,
+      "returnAvsData" : false,
+      "includeWeChatPayOpenid" : false,
+      "includeCustomRoutingFlagging" : false,
+      "includeTokenisedPaymentMetaData" : false
+    }
+  },
+  "_links" : {
+    "self" : {
+      "href" : "https://management-test.adyen.com/v1/merchants/YOUR_MERCHANT_ACCOUNT/webhooks/S2-31433949437F"
+    },
+    "generateHmac" : {
+      "href" : "https://management-test.adyen.com/v1/merchants/YOUR_MERCHANT_ACCOUNT/webhooks/S2-31433949437F/generateHmac"
+    },
+    "merchant" : {
+      "href" : "https://management-test.adyen.com/v1/merchants/YOUR_MERCHANT_ACCOUNT"
+    },
+    "testWebhook" : {
+      "href" : "https://management-test.adyen.com/v1/merchants/YOUR_MERCHANT_ACCOUNT/webhooks/S2-31433949437F/test"
+    }
+  }
+}
