+++ tests/Resources/Management/post-merchantIs-add-allowed-origin-success.json
@@ -1,0 +1,3 @@
+{
+  "domain" : "https://test-website.com/"
+}
