+++ tests/Resources/Management/post-merchants-webhooks-generateHmac-success.json
@@ -1,0 +1,3 @@
+{
+  "hmacKey" : "AG4345BV33J78K17F4F963516FA42AC8813EWS3J8F41E902"
+}
